refactor: move component into subdir
