+++ src/page/Dashboard/DashboardApps/index.tsx
@@ -49,8 +49,8 @@ export const DashboardApps: FC = () => {
           dashboardAppActions.updateDashboardAppListReducer(response.data),
         )
       },
-      (response) => { },
-      (error) => { },
+      (response) => {},
+      (error) => {},
       (loading) => {
         setLoading(loading)
       },
@@ -83,8 +83,8 @@ export const DashboardApps: FC = () => {
                 )
                 navigate(`/editor/${response.data.appId}`)
               },
-              (response) => { },
-              (error) => { },
+              (response) => {},
+              (error) => {},
               (loading) => {
                 setCreateNewLoading(loading)
               },
