+++ src/page/Editor/components/ActionEditor/ActionEditorPanel/ResourcePanel/interface.ts
@@ -1,0 +1,6 @@
+export interface ResourcePanelProps {
+  resourceId?: string
+  activeActionItemId?: string | null
+  onChange?: () => void
+  onSave?: () => void
+}
