+++ src/page/Editor/components/ActionEditor/ActionEditorPanel/TitleInput/interface.ts
@@ -1,0 +1,5 @@
+export interface TitleInputProps
+  extends Omit<HTMLAttributes<HTMLDataElement>, "title"> {
+  title?: string
+  activeActionItem?: ActionItem | null
+}
