+++ src/page/Editor/components/ActionEditor/ActionList/style.tsx
@@ -35,8 +35,8 @@ export const applynewButtonStyle = (isActive: boolean): SerializedStyles => {
 
     &:hover {
       background-color: ${globalColor(
-    `--${illaPrefix}-techPurple-06`,
-  )}!important;
+        `--${illaPrefix}-techPurple-06`,
+      )}!important;
     }
   `
 }
@@ -251,8 +251,8 @@ export const searchInputStyle = css`
     border-color: ${globalColor(`--${illaPrefix}-techPurple-01`)}!important;
     box-shadow: 0 0 8px 0
       ${chroma(globalColor(`--${illaPrefix}-techPurple-01`))
-    .alpha(0.2)
-    .hex()};
+        .alpha(0.2)
+        .hex()};
   }
 `
